requirements.txt
Flask==3.0.3
google-cloud-firestore==2.18.0 
google-cloud-storage==2.18.2
pytest==8.3.3
python-dotenv==1.0.1